Major updates to contact page.
Contact Page: Removed <table> and used <ul> to properly set up the links to social media. <table> made the positioning of the links easier, but the links wouldn't work correctly. Added an id to the top navigation bar (Only added on the contact page), for better targeting of other <ul> inside the top div. Commented out the javascript that was used for submitting the form:Not working as intended/not properlly used.

Gerneral: Finished adding the URLs to all the links in the footer section of each page. Changed the width declaration to percentage% and added a min&max width. Added meta data "keyword" & "description"  to all pages.